+++ package.json
@@ -88,6 +88,7 @@
     "prettier": "^1.14.3",
     "redux-mock-store": "^1.5.3",
     "rimraf": "^2.6.2",
+    "supertest": "^3.3.0",
     "webpack": "^4.19.1",
     "webpack-cli": "^3.1.1",
     "webpack-dev-server": "^3.1.8"
