--- src/server.test.js
@@ -1,1 +1,0 @@
-test('test', () => {})
