+++ src/server/server.test.js
@@ -1,0 +1,9 @@
+const request = require('supertest')
+const app = require('./app')
+
+describe('Test the root path', () => {
+  test('It should response the GET method', async () => {
+    const response = await request(app).get('/')
+    expect(response.statusCode).toBe(200)
+  })
+})
